+++ client/coral-admin/src/AppRouter.js
@@ -5,7 +5,7 @@ import ModerationQueue from 'containers/ModerationQueue'
 import CommentStream from 'containers/CommentStream'
 import EmbedLink from 'components/EmbedLink'
 import Configure from 'containers/Configure'
-import CommunityContainer from 'containers/CommunityContainer'
+import CommunityContainer from 'containers/Community/CommunityContainer'
 import LayoutContainer from 'containers/LayoutContainer'
 
 const routes = (
