+++ client/coral-admin/src/containers/Community/Community.js
@@ -1,0 +1,42 @@
+import React, { Component } from 'react'
+import { connect } from 'react-redux'
+import I18n from 'coral-framework/i18n/i18n'
+import translations from '../../translations'
+import { Grid, Cell, Button } from 'react-mdl'
+
+import styles from './Community.css'
+import Table from './Table'
+
+export default class Community extends Component {
+  render() {
+    const {searchValue, onKeyDownHandler, onChangeHandler} = this.props;
+    return (
+        <Grid>
+          <Cell col={4}>
+            <input
+              className={styles.searchInput}
+              type="text"
+              value={searchValue}
+              onKeyDown={onKeyDownHandler}
+              onChange={onChangeHandler}
+            />
+            <button className={styles.roleButton}>
+              All Users
+            </button>
+          </Cell>
+          <Cell col={8}>
+            <Table
+              headers={['Username and Email', 'Account Creation Date']}
+              data={[
+                {
+                  name: 'Belen Curcio',
+                  email: 'curciobelen@gmail.com',
+                  date: '21/12/2010'
+                }
+              ]}
+            />
+          </Cell>
+        </Grid>
+    )
+  }
+}
